+++ Models/EncodeFinalization.cs
@@ -32,6 +32,7 @@ namespace MediaFlux.Models
         };
         public MediaProbeResult? SourceProbe { get; init; }
         public double? ExpectedDurationSeconds { get; init; }
+        public long? ExpectedVideoFrameCount { get; init; }
         public int? ExpectedVideoWidth { get; init; }
         public int? ExpectedVideoHeight { get; init; }
         public PerformanceTimingService? PerformanceTiming { get; init; }